Merged changes from google version:  cl/130981487
Created a manual binding for glVertexAttribPointer(), so it would work properly.  The generated version used a TypedData argument for the offset, so there was no way to specify a valid offset within the vertex data.  That argument is now an integer (number of bytes for offset).  The example_triangle demo was updated to show color values interleaved with vertex values.